Fix input function and change from type int to size_t

+++ zubarev.arsenii/S1/data.hpp
@@ -10,7 +10,7 @@ namespace zubarev
   struct Data
   {
     std::string name;
-    List< int > numbers;
+    List< size_t > numbers;
   };
 }
